consumer-app/requirements.txt
kafka-python==2.0.2
elasticsearch==8.6.2
geoip2==4.7.0
requests==2.31.0
